+++ Internship2022TeamPatri/Internship2022TeamPatri/Views/Home/Profile/ProfileViewController.xib
@@ -7,7 +7,7 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="PProfileViewController" customModule="Internship2022TeamPatri" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ProfileViewController">
             <connections>
                 <outlet property="emailLabel" destination="ktT-ah-5d2" id="PIO-yr-1lQ"/>
                 <outlet property="emailTextLabel" destination="x9k-Ri-0VB" id="F4c-gL-aFh"/>
